WIP tests

+++ Task_2_2_1/src/main/java/Config.java
@@ -15,6 +15,6 @@ record Config(
         int bakers,
         int couriers,
         int warehouseCapacity,
-        int[] bakersSpeed,
-        int[] couriersCapacity
+        long[] bakersSpeed,
+        long[] couriersCapacity
 ) {}
